Split open popup functionality into separate file.
